+++ includes/admin/class.llms.admin.notices.core.php
@@ -3,7 +3,7 @@
  * Manage core admin notices
  *
  * @since    3.0.0
- * @version  3.14.7
+ * @version  [version]
  */
 
 if ( ! defined( 'ABSPATH' ) ) { exit; }
@@ -13,9 +13,12 @@ class LLMS_Admin_Notices_Core {
 	/**
 	 * Costructor
 	 * @since    3.0.0
-	 * @version  3.0.0
+	 * @version  [version]
 	 */
 	public static function init() {
+
+		add_action( 'admin_head', array( __CLASS__, 'maybe_hide_notices' ), 1 );
+		add_action( 'current_screen', array( __CLASS__, 'maybe_hide_notices' ), 999 );
 
 		add_action( 'current_screen', array( __CLASS__, 'add_init_actions' ) );
 		add_action( 'switch_theme', array( __CLASS__, 'clear_sidebar_notice' ) );
@@ -122,6 +125,24 @@ class LLMS_Admin_Notices_Core {
 		}
 	}
 
+	/**
+	 * Don't display notices on specific pages
+	 * @return   void
+	 * @since    [version]
+	 * @version  [version]
+	 */
+	public static function maybe_hide_notices() {
+
+		$screen = get_current_screen();
+
+		if ( 'admin_page_llms-course-builder' === $screen->id ) {
+
+			remove_all_actions( 'admin_notices' ); // 3rd party notices
+			remove_action( 'admin_print_styles', array( 'LLMS_Admin_Notices', 'output_notices' ) ); // notices output by LifterLMS
+
+		}
+
+	}
 
 	/**
 	 * Check theme support for LifterLMS Sidebars